Rename newStripeCheckout to stripeCheckout
