+++ assets/helpers/paymentIntegrations/newPaymentFlow/stripeCheckout.js
